Wallet api tweaks

+++ src/WalletBackend/WalletBackend.cpp
@@ -252,6 +252,11 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
         privateSpendKey, privateViewKey
     );
 
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
     /* Just defining here so it's more obvious what we're doing in the
        constructor */
     bool newWallet = false;
@@ -286,6 +291,16 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
         return {error, nullptr}; 
     }
 
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
     /* Just defining here so it's more obvious what we're doing in the
        constructor */
     bool newWallet = false;
@@ -318,6 +333,18 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importViewWalle
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
         return {error, nullptr}; 
+    }
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, nullptr};
     }
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
@@ -763,6 +790,11 @@ std::tuple<Error, std::string> WalletBackend::importSubWallet(
     const Crypto::SecretKey privateSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importSubWallet(
@@ -794,6 +826,11 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
     const Crypto::PublicKey publicSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePublicKey(publicSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importViewSubWallet(
@@ -823,6 +860,13 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
 
 Error WalletBackend::deleteSubWallet(const std::string address)
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return error;
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         return m_subWallets->deleteSubWallet(address);
     });
@@ -889,6 +933,13 @@ Error WalletBackend::changePassword(const std::string newPassword)
 std::tuple<Error, Crypto::PublicKey, Crypto::SecretKey>
     WalletBackend::getSpendKeys(const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, Crypto::PublicKey(), Crypto::SecretKey()};
+    }
+
     const auto [publicSpendKey, publicViewKey] = Utilities::addressToKeys(address);
 
     const auto [success, privateSpendKey] = m_subWallets->getPrivateSpendKey(publicSpendKey);
@@ -915,6 +966,13 @@ std::tuple<Error, std::string> WalletBackend::getMnemonicSeed() const
 std::tuple<Error, std::string> WalletBackend::getMnemonicSeedForAddress(
     const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, std::string()}; 
+    }
+
     const auto privateViewKey = getPrivateViewKey();
     const auto [error, publicSpendKey, privateSpendKey] = getSpendKeys(address);
 
@@ -1016,6 +1074,11 @@ bool WalletBackend::daemonOnline() const
 std::tuple<Error, std::string> WalletBackend::getAddress(
     const Crypto::PublicKey spendKey) const
 {
+    if (Error error = validatePublicKey(spendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_subWallets->getAddress(spendKey);
 }
